code/requirements.txt
# Function dependencies, for example:
#package>=version
flask
geopy
requests
google-cloud-storage